FingerTip Banking App
I worked on a project where I used C++ advanced programming techniques to create a banking system, that was to solve modern problems. The banking system allowed customers to open bank accounts on the system and use their virtual cards received upon opening bank accounts to pay for items online or in-store. The aim was to reduce the number of customers who went to branches to open bank accounts during the pandemic.